Split generic behavior extends diagnostics tests
